refactor: #13 DDDに切り替えようとしたが冗長過ぎるので中止
