+++ electron-builder.yml
@@ -1,4 +1,4 @@
-appId: com.electron.app
+appId: dev.minr.desktop
 productName: minr-desktop
 directories:
   buildResources: build
@@ -13,6 +13,7 @@ asarUnpack:
   - resources/**
 afterSign: build/notarize.js
 win:
+  icon: build/icon.ico
   executableName: minr-desktop
 nsis:
   artifactName: ${name}-${version}-setup.${ext}
@@ -20,6 +21,7 @@ nsis:
   uninstallDisplayName: ${productName}
   createDesktopShortcut: always
 mac:
+  icon: build/icon.icns
   entitlementsInherit: build/entitlements.mac.plist
   extendInfo:
     - NSCameraUsageDescription: Application requests access to the device's camera.
@@ -29,6 +31,7 @@ mac:
 dmg:
   artifactName: ${name}-${version}.${ext}
 linux:
+  icon: build/icon.png
   target:
     - AppImage
     - snap
