+++ src/main/ipc/CredentialsStoreServiceHandlerImpl.ts
