+++ src/main/ipc/GoogleAuthServiceHandlerImpl.ts
